#49 feat: refactor solution DTO structure
