+++ src/main/java/com/example/algoproject/solution/dto/request/AddSolution.java
@@ -1,4 +1,4 @@
-package com.example.algoproject.solution.dto;
+package com.example.algoproject.solution.dto.request;
 
 import lombok.Data;
 import org.springframework.web.multipart.MultipartFile;
