+++ src/main/java/com/example/algoproject/solution/service/SolutionService.java
@@ -10,9 +10,10 @@ import com.example.algoproject.problem.domain.Problem;
 import com.example.algoproject.problem.repository.ProblemRepository;
 import com.example.algoproject.s3.S3Uploader;
 import com.example.algoproject.solution.domain.Solution;
-import com.example.algoproject.solution.dto.AddSolution;
-import com.example.algoproject.solution.dto.CommitFileRequest;
-import com.example.algoproject.solution.dto.S3UrlResponse;
+import com.example.algoproject.solution.dto.request.AddSolution;
+import com.example.algoproject.solution.dto.request.CommitFileRequest;
+import com.example.algoproject.solution.dto.request.UpdateSolution;
+import com.example.algoproject.solution.dto.response.SolutionInfo;
 import com.example.algoproject.solution.repository.SolutionRepository;
 import com.example.algoproject.study.domain.Study;
 import com.example.algoproject.study.repository.StudyRepository;
